refactor: generic sync framework, unified DB access, consolidated interfaces
- Add generic sync framework (internal/sync/) with Syncer[T] interface,
  refactored RPC SyncProcedures as proof of concept (-85 lines)
- Unify database access: AuthHandler, OAuthHandler, SQLHandler,
  ServiceKeyHandler now use *database.Connection instead of *pgxpool.Pool
- Delegate api/rest_errors.go to shared errors/http.go (36 constants,
  16 Send* functions now thin wrappers over apperrors.*)
- Adopt context helpers across functions, jobs, rpc, secrets, ai handlers
  (replacing raw c.Locals() calls with middleware.GetUserID etc.)
- Consolidate 5 auth email interfaces into single EmailService interface
- Split database/connection.go into connection.go + connection_metrics.go
- Split auth/service.go into service.go + service_users.go + service_tokens.go
- Split functions/handler.go into handler.go + handler_execute.go + handler_sync.go
- Split jobs/handler.go into handler.go + handler_jobs.go + handler_sync.go
- Add nil-safe Connection.Pool() for test compatibility
- All tests pass, linter clean
