+++ internal/database/connection.go
@@ -7,7 +7,6 @@ import (
 	"os"
 	"path/filepath"
 	"regexp"
-	"runtime"
 	"sort"
 	"strings"
 	"sync"
@@ -21,7 +20,6 @@ import (
 	"github.com/rs/zerolog/log"
 
 	"github.com/nimbleflux/fluxbase/internal/config"
-	"github.com/nimbleflux/fluxbase/internal/logutil"
 	"github.com/nimbleflux/fluxbase/internal/observability"
 )
 
@@ -30,79 +28,6 @@ type callerKey struct{}
 func WithCaller(ctx context.Context, caller string) context.Context {
 	return context.WithValue(ctx, callerKey{}, caller)
 }
-
-func getCallerFromContext(ctx context.Context) string {
-	if ctx == nil {
-		return ""
-	}
-	if v, ok := ctx.Value(callerKey{}).(string); ok && v != "" {
-		return v
-	}
-	return ""
-}
-
-func getCallerFromRuntime() string {
-	for skip := 3; skip <= 8; skip++ {
-		if _, file, _, ok := runtime.Caller(skip); ok {
-			idx := strings.LastIndex(file, "/internal/")
-			if idx >= 0 {
-				return file[idx+1:]
-			}
-		}
-	}
-	return ""
-}
-
-type slowQueryEntry struct {
-	count     int
-	firstSeen time.Time
-}
-
-type slowQueryTracker struct {
-	mu      sync.Mutex
-	entries map[string]*slowQueryEntry
-	maxAge  time.Duration
-}
-
-func newSlowQueryTracker() *slowQueryTracker {
-	t := &slowQueryTracker{
-		entries: make(map[string]*slowQueryEntry),
-		maxAge:  1 * time.Hour,
-	}
-	go t.cleanupLoop()
-	return t
-}
-
-func (t *slowQueryTracker) record(queryKey string) int {
-	t.mu.Lock()
-	defer t.mu.Unlock()
-
-	now := time.Now()
-	e, ok := t.entries[queryKey]
-	if !ok {
-		t.entries[queryKey] = &slowQueryEntry{count: 1, firstSeen: now}
-		return 1
-	}
-	e.count++
-	return e.count
-}
-
-func (t *slowQueryTracker) cleanupLoop() {
-	ticker := time.NewTicker(10 * time.Minute)
-	defer ticker.Stop()
-	for range ticker.C {
-		t.mu.Lock()
-		now := time.Now()
-		for k, e := range t.entries {
-			if now.Sub(e.firstSeen) > t.maxAge {
-				delete(t.entries, k)
-			}
-		}
-		t.mu.Unlock()
-	}
-}
-
-const slowQueryTruncationLimit = 500
 
 // These aliases allow the middleware and handlers to use simpler type names
 type (
@@ -130,11 +55,6 @@ type Connection struct {
 	metrics            *observability.Metrics
 	slowQueryTracker   *slowQueryTracker
 	slowQueryThreshold time.Duration
-}
-
-// SetMetrics sets the metrics instance for recording database metrics
-func (c *Connection) SetMetrics(m *observability.Metrics) {
-	c.metrics = m
 }
 
 // ExtractTableName attempts to extract the table name from a SQL query
@@ -325,9 +245,18 @@ func (c *Connection) Close() {
 
 // Pool returns the underlying connection pool
 func (c *Connection) Pool() *pgxpool.Pool {
+	if c == nil {
+		return nil
+	}
 	c.poolMu.RLock()
 	defer c.poolMu.RUnlock()
 	return c.pool
+}
+
+// NewConnectionFromPool creates a minimal Connection wrapping an existing pool.
+// Intended for tests only.
+func NewConnectionFromPool(pool *pgxpool.Pool) *Connection {
+	return &Connection{pool: pool}
 }
 
 // RecreatePool closes the current pool and creates a new one.
@@ -848,50 +777,6 @@ func (c *Connection) Stats() *pgxpool.Stat {
 		return nil
 	}
 	return c.pool.Stat()
-}
-
-func (c *Connection) logSlowQuery(ctx context.Context, sql string, duration time.Duration, opType string) {
-	if duration <= c.slowQueryThreshold {
-		return
-	}
-
-	operation := ExtractOperation(sql)
-	table := ExtractTableName(sql)
-	sanitizedQuery := truncateQuery(logutil.SanitizeSQL(sql), slowQueryTruncationLimit)
-
-	queryKey := operation + ":" + table
-	occurrences := 1
-	if c.slowQueryTracker != nil {
-		occurrences = c.slowQueryTracker.record(queryKey)
-	}
-
-	caller := getCallerFromContext(ctx)
-	if caller == "" {
-		caller = getCallerFromRuntime()
-	}
-
-	evt := log.Warn().
-		Dur("duration", duration).
-		Int64("duration_ms", duration.Milliseconds()).
-		Str("operation", operation).
-		Str("table", table).
-		Str("query", sanitizedQuery).
-		Int("occurrences", occurrences).
-		Bool("slow_query", true)
-
-	if caller != "" {
-		evt = evt.Str("caller", caller)
-	}
-
-	evt.Msg("Slow query detected")
-}
-
-// truncateQuery truncates a SQL query to a maximum length for logging
-func truncateQuery(query string, maxLen int) string {
-	if len(query) <= maxLen {
-		return query
-	}
-	return query[:maxLen] + "... (truncated)"
 }
 
 // WrapWithServiceRole wraps a database operation with service_role context
